requirements.txt
-e .
aws-cdk.aws-ec2
aws-cdk.aws-elasticloadbalancingv2